Move trace and connection to own files.
