+++ Fabric/Graph/GraphRenderer.swift
@@ -9,58 +9,6 @@ import SwiftUI
 import Metal
 import Satin
 
-private final class GraphExecutionTraceFrameBuilder
-{
-    let executionIndex: Int
-    let startedAt: TimeInterval
-    var nodeExecutions: [NodeExecution] = []
-
-    init(executionIndex: Int, startedAt: TimeInterval)
-    {
-        self.executionIndex = executionIndex
-        self.startedAt = startedAt
-    }
-
-    func makeExecution(endedAt: TimeInterval) -> GraphExecution
-    {
-        GraphExecution(executionIndex: executionIndex,
-                       startedAt: startedAt,
-                       endedAt: endedAt,
-                       nodeExecutions: nodeExecutions)
-    }
-}
-
-private final class NodeExecutionTraceBuilder
-{
-    let nodeID: UUID
-    let nodeName: String
-    let nodeTypeName: String
-    let orderIndex: Int
-    let startedAt: TimeInterval
-    var childExecutions: [GraphExecution] = []
-
-    init(node: Node, orderIndex: Int, startedAt: TimeInterval)
-    {
-        self.nodeID = node.id
-        self.nodeName = node.name
-        self.nodeTypeName = String(describing: type(of: node))
-        self.orderIndex = orderIndex
-        self.startedAt = startedAt
-    }
-
-    func makeNodeExecution(endedAt: TimeInterval, result: NodeExecutionResult) -> NodeExecution
-    {
-        NodeExecution(nodeID: nodeID,
-                      nodeName: nodeName,
-                      nodeTypeName: nodeTypeName,
-                      orderIndex: orderIndex,
-                      startedAt: startedAt,
-                      endedAt: endedAt,
-                      result: result,
-                      childExecutions: childExecutions)
-    }
-}
-
 // Graph Execution Engine
 public class GraphRenderer : ViewRenderer
 {
